+++ ui/CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to the **Prowler UI** are documented in this file.
 ## [v1.9.0] (Prowler v5.9.0) – UNRELEASED
 
 ### 🚀 Added
+
+- Unhid SAML-related components to make them available [(#8203)](https://github.com/prowler-cloud/prowler/pull/8203)
+
 ### 🔄 Changed
 
 - Upgrade to Next.js 14.2.30 and lock TypeScript to 5.5.4 for ESLint compatibility [(#8189)](https://github.com/prowler-cloud/prowler/pull/8189)