+++ ui/CHANGELOG.md
@@ -14,6 +14,14 @@ All notable changes to the **Prowler UI** are documented in this file.
 - Upgrade to Next.js 14.2.30 and lock TypeScript to 5.5.4 for ESLint compatibility [(#8189)](https://github.com/prowler-cloud/prowler/pull/8189)
 
 ### 🐞 Fixed
+### Removed
+
+---
+
+## [v1.8.1] (Prowler 5.8.1)
+
+### Removed
+- Validation of the provider's secret type during updates [(#8197)](https://github.com/prowler-cloud/prowler/pull/8197)
 
 ---